split disp tokens code into separate files, mimicking disptokens
