+++ radiant/disp.h
@@ -199,50 +199,3 @@ class Disp final:
 
     // TODO: (quad, power, DispVertArray[index]) -> DispRenderVert(bary.pos + offset, bary.uv, blend)
 };
-
-
-class DispTokenImporter: public MapImporter {
-  private:
-    Disp &m_disp;
-
-    inline static bool importHeader(Disp &disp, Tokeniser &tokeniser);
-    inline static bool importShader(Disp &disp, Tokeniser &tokeniser);
-    inline static bool importParams(Disp &disp, Tokeniser &tokeniser);
-    inline static bool importVertex(Disp &disp, Tokeniser &tokeniser);
-    inline static bool importFooter(Disp &disp, Tokeniser &tokeniser);
-
-  public:
-    PatchTokenImporter(Disp &disp) : m_disp(disp) {};
-
-    bool importTokens(Tokeniser &tokeniser) override {
-        RETURN_FALSE_IF_FAIL(Disp_importHeader(m_disp, tokeniser));
-        RETURN_FALSE_IF_FAIL(Disp_importShader(m_disp, tokeniser));
-        RETURN_FALSE_IF_FAIL(Disp_importParams(m_disp, tokeniser));
-        RETURN_FALSE_IF_FAIL(Disp_importVertex(m_disp, tokeniser));
-        RETURN_FALSE_IF_FAIL(Disp_importFooter(m_disp, tokeniser));
-        return true;
-    }
-};
-
-
-class DispTokenExporter : public MapExporter {
-  private:
-    const Disp &m_disp;
-
-    inline static void exportHeader(Disp &disp, TokenWriter &writer);
-    inline static void exportShader(Disp &disp, TokenWriter &writer);
-    inline static void exportParams(Disp &disp, TokenWriter &writer);
-    inline static void exportVertex(Disp &disp, TokenWriter &writer);
-    inline static void exportFooter(Disp &disp, TokenWriter &writer);
-
-  public:
-    PatchTokenExporter(Disp &disp) : m_disp(disp) {}
-
-    void exportTokens(TokenWriter &writer) const override {
-        exportHeader(m_disp, writer);
-        exportShader(m_disp, writer);
-        exportParams(m_disp, writer);
-        exportVertex(m_disp, writer);
-        exportFooter(m_disp, writer);
-    }
-};
